refactor: cookie functions moved to one file
