+++ src/utils/index.ts
@@ -1,7 +1,5 @@
 import { getDomainFromHostname } from './getDomainFromHostname'
-import { createCookieStringFromObject } from './createCookieStringFromObject'
 import { getCacheControlHeaderWithMaxAgeIfLower } from './getCacheControlHeaderWithMaxAgeIfLower'
-import { createCookieObjectFromHeaderValue } from './createCookieObjectFromHeaderValue'
 import { createErrorResponse } from './createErrorResponse'
 import { fetchCacheable } from './fetchCacheable'
 import {
@@ -9,7 +7,7 @@ import {
   addTrafficMonitoringSearchParamsForProCDN,
 } from './addTrafficMonitoring'
 import { returnHttpResponse } from './returnHttpResponse'
-import { Cookie } from './cookie'
+import { Cookie, createCookieStringFromObject, createCookieObjectFromHeaderValue } from './cookie'
 
 export {
   getDomainFromHostname,
